requirements.txt
attrs==21.4.0
Authlib==0.15.5
cached-property==1.5.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
idna==3.3
isodate==0.6.1
lxml==4.8.0
pkg_resources==0.0.0
platformdirs==2.5.1
pycparser==2.21
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
requests-file==1.5.1
requests-toolbelt==0.9.1
simple-salesforce==1.11.5
six==1.16.0
urllib3==1.26.8
zeep==4.1.0
